+++ src/apis/markdown/fetch-contents.ts
@@ -1,4 +1,3 @@
-// "use server"; // MEMO: commented because functions are not async functions.
 import "server-only";
 import fs from "node:fs";
 import { join } from "node:path";
