+++ src/features/dump/actions/add-contents.ts
@@ -3,7 +3,8 @@ import "server-only";
 import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { createSelfContents } from "@/apis/prisma/fetch-contents";
 import { SUCCESS_MESSAGES } from "@/constants";
-import { NotAllowedError, wrapServerSideErrorForClient } from "@/error";
+import { NotAllowedError } from "@/error-classes";
+import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { checkPostPermission } from "@/features/auth/utils/role";
 import type { ContentsContext } from "@/features/dump/stores/contents-context";
 import { validateContents } from "@/features/dump/utils/validate-contents";
