+++ src/features/dump/actions/add-news.ts
@@ -4,7 +4,8 @@ import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { createSelfCategory } from "@/apis/prisma/fetch-category";
 import { createSelfNews } from "@/apis/prisma/fetch-news";
 import { SUCCESS_MESSAGES } from "@/constants";
-import { NotAllowedError, wrapServerSideErrorForClient } from "@/error";
+import { NotAllowedError } from "@/error-classes";
+import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { checkPostPermission } from "@/features/auth/utils/role";
 import type { NewsContext } from "@/features/dump/stores/news-context";
 import { validateCategory } from "@/features/dump/utils/validate-category";
