+++ src/features/profile/actions/change-profile.ts
@@ -3,7 +3,7 @@ import "server-only";
 import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { upsertSelfProfile } from "@/apis/prisma/fetch-profile";
 import { SUCCESS_MESSAGES } from "@/constants";
-import { wrapServerSideErrorForClient } from "@/error";
+import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import type { ProfileSchema } from "@/features/profile/schemas/profile-schema";
 import type { ServerAction } from "@/types";
 import { formatUpsertProfileMessage } from "@/utils/format-for-line";
