+++ tsconfig.json
@@ -18,5 +18,5 @@
 		"paths": { "@/*": ["./src/*"] }
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-	"exclude": ["node_modules", "script"]
+	"exclude": ["node_modules", "script/**/*", "README.md"]
 }
